clean up benchmark
